requirements.txt
requests
openai
python-dotenv
fastapi
uvicorn
pydantic
python-dotenv
pytesseract
pillow
pandas
deep-translator